fix: checked flash address parse, builders in flash_args module
Review feedback (CodeRabbit):
- parse_flash_u32 rejects values past u32 (deviation: C's strtoul
  truncation made -a 0x100000000 address 0 — with a forced write, an
  erase plus firmware written at the wrong address). rx keeps its
  faithful truncating parse_u32 for the C rx tool's semantics.
  Regression test added.
- The clap builders move to their own flash_args module, restoring
  flash_cli's tests-at-bottom layout and dropping the
  items_after_test_module suppression. Bonus: Lizard now measures
  every function in both files correctly (verified locally).
